+++ Odotocodot.OneNote.Linq/Internal/IWritableHasPath.cs
@@ -1,9 +1,0 @@
-using Odotocodot.OneNote.Linq.Abstractions;
-
-namespace Odotocodot.OneNote.Linq.Internal
-{
-    internal interface IWritableHasPath : IHasPath
-    {
-        new string Path { set; }
-    }
-}