+++ src/LinqToOneNote/Abstractions/IHasColor.cs
@@ -1,0 +1,17 @@
+using System.Drawing;
+
+namespace LinqToOneNote.Abstractions
+{
+	/// <summary>
+	/// Represents an OneNote hierarchy item that has a color.
+	/// </summary>
+	/// <seealso cref="Notebook"/>
+	/// <seealso cref="Section"/>
+	public interface IHasColor : IOneNoteItem
+	{
+		/// <summary>
+		/// The color of the notebook.
+		/// </summary>
+		Color? Color { get; }
+	}
+}