+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [2.0.0] -
+
+- Updated to .Net 9.0
+- Renamed ID to Id
+- Renamed classes - Removed OneNote from names OneNoteNotebook.
+- (Temporarily?) Removed RelativePath and Notebook Properties
+- Refactor Name Validation
+- Added support for OpenSections and UnfiledNotes 
+- Add support for get a partial hierarchy
+- Rename Section property to Parent on page
+- Rename Traverse to Descendants
+- Add alternative methods for getting a com object
+- Rename OneNoteApplication to OneNote
+
+
 ## [1.2.0] - 2025-07-05
 
 ### What’s Changed